% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigvaldispRcpp-package.R
\docType{package}
\name{eigvaldispRcpp-package}
\alias{eigvaldispRcpp-package}
\alias{eigvaldispRcpp}
\title{eigvaldispRcpp: Rcpp Extension for 'eigvaldisp'}
\description{
This package extends the main package \pkg{eigvaldisp} with
fast \proglang{C++} functions via the \pkg{Rcpp} interface
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{eigvaldispRcpp}
\packageIndices{eigvaldispRcpp}
}
\section{Author/Maintainer}{

Junya Watanabe \href{mailto:jw2098@cam.ac.uk}{jw2098@cam.ac.uk}
}

\examples{
## Generate a (moderately) large population correlation matrix
Rho <- GenCov(p = 50L, shape = "linearly_decreasing", evectors = "Givens")

## Arbitrary sample sizes
Ns <- c(10, 20, 50)

## Calculate the variance of Vrel(Rho)
## When the argument cppfun is provided, Var.VRR() automatically calls
## eigvaldisp:::AVar.VRR_pfc():
eigvaldisp::Var.VRR(Rho, Ns - 1, cppfun = "Cov_r2C") # Base Rcpp
eigvaldisp::Var.VRR(Rho, Ns - 1, cppfun = "Cov_r2A") # RcppArmadillo + OpenMP
eigvaldisp::Var.VRR(Rho, Ns - 1, cppfun = "Cov_r2E") # RcppEigen + OpenMP
eigvaldisp::Var.VRR(Rho, Ns - 1, cppfun = "Cov_r2P") # RcppParallel
## These use four alternative functions provided and yield identical results
## The last three parallelize the computation when possible
## Speed may vary, depending on environment

}
\references{
Watanabe, J. (2021) eigvaldisp: R package for statistics of
eigenvalue dispersion.
\url{https://github.com/watanabe-j/eigvaldisp}.

Watanabe, J. (2022) Statistics of eigenvalue dispersion indices:
quantifying the magnitude of phenotypic integration. \emph{Evolution},
\strong{76}, 4--28. \doi{10.1111/evo.14382}.
}
\seealso{
\code{\link[eigvaldisp:eigvaldisp-package]{eigvaldisp}}: The main package

\code{\link{Cov_r2}}: The \proglang{C++} functions

\code{\link[eigvaldisp]{Var.VRR}}, \code{\link[eigvaldisp]{AVar.VRR_pfc}}:
Outer functions in which the \proglang{C++} functions are internally used
}
